edit submit page

+++ Possible/src/main/java/com/yedam/possable/app/rent/controller/PremiumRentController.java
@@ -15,6 +15,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.HttpServletRequest;
 import java.security.Principal;
 import java.util.*;
 
@@ -57,16 +58,10 @@ public class PremiumRentController {
     public String estimateInsert(EstimateHistoryVO vo,
                                      @RequestParam("options") String[] optionsArr,
                                      @RequestParam("items") String[] itemsArr,
-                                     @RequestParam("memSeq") Long memSeq,
                                      RedirectAttributes attributes) {
         // 옵션 배열 -> 스트링
         vo.setOptions(Arrays.toString(optionsArr));
         vo.setItems(Arrays.toString(itemsArr));
-
-        // 외래키 객체 설정
-        MemberVO memVo = new MemberVO();
-        memVo.setSeq(memSeq);
-        vo.setMemberVO(memVo);
 
         // 코드 -> 네임 변환
         vo.setSegment(codeService.getNameByCode(vo.getSegment()).getName());
@@ -90,8 +85,20 @@ public class PremiumRentController {
     // 견적 요청 상세
     @GetMapping("estimate/view")
     public String estimateRead(@RequestParam Long seq,
+                               Principal principal,
+                               RedirectAttributes attributes,
+                               HttpServletRequest request,
                                Model model,
                                @ModelAttribute Criteria cri) {
+        MemberVO user = (MemberVO) principal;
+
+//        if(user == null || user.getAuthor().equals("USER")){
+//            String denyMsg = "업체회원만 열람 가능합니다.";
+//            attributes.addFlashAttribute("denyMsg", denyMsg);
+//
+//            return "redirect:" + request.getHeader("REFERER");
+//        }
+
         model.addAttribute("estimate", premiumRentService.getEstimate(seq));
 
         return "/premiumRent/estimateView";
@@ -101,8 +108,18 @@ public class PremiumRentController {
     @GetMapping("estimate/delete")
     public String estimateDelete(@RequestParam Long seq,
                                  RedirectAttributes attributes,
+                                 HttpServletRequest request,
                                  Principal principal){
         // principal 객체로 견적 소유자 검증(id or seq)
+        EstimateHistoryVO vo = (EstimateHistoryVO) premiumRentService.getEstimate(seq).get("estimate");
+        String user = ((MemberVO)principal).getId();
+        String writer = vo.getMemSeq().toString();
+
+        if(user == null || !user.equals(writer)){
+            attributes.addFlashAttribute("deleteMsg", "작성자만 삭제 가능합니다.");
+            return "redirect:" + request.getHeader("REFERER");
+        }
+
         int deleteResult = premiumRentService.deleteEstimate(seq);
 
         String resultMsg = "";
@@ -120,8 +137,18 @@ public class PremiumRentController {
     @GetMapping("estimate/update")
     public String estimateUpdateForm(@RequestParam Long seq,
                                      Principal principal,
+                                     HttpServletRequest request,
+                                     RedirectAttributes attributes,
                                      Model model){
         // principal 객체로 견적 소유자 검증(id or seq)
+        EstimateHistoryVO vo = (EstimateHistoryVO) premiumRentService.getEstimate(seq).get("estimate");
+        String user = ((MemberVO)principal).getId();
+        String writer = vo.getMemSeq().toString();
+
+        if(user == null || !user.equals(writer)){
+            attributes.addFlashAttribute("updateMsg", "작성자만 수정 가능합니다.");
+            return "redirect:" + request.getHeader("REFERER");
+        }
         model.addAttribute("estimate", premiumRentService.getEstimate(seq));
 
         return "premiumRent/estimateInsert";
@@ -145,19 +172,20 @@ public class PremiumRentController {
         return "redirect: view";
     }
 
+    // 견적 작성
     @GetMapping("submit/insert")
     public String submitInsert(Model model,
-                               @RequestParam("estSeq") Long estSeq,
+                               @RequestParam Long seq,
                                Principal principal){
-        EstimateHistoryVO estimateHistoryVO = premiumRentService.getEstimate(estSeq);
-        Long companySeq = 1L; // principal.getName();    pricipal에서 업체 시퀀스 조회
+        Map<String, Object> estimate = premiumRentService.getEstimate(seq);
+        Long companySeq = ((MemberVO)principal).getSeq(); // principal.getName();    pricipal에서 업체 시퀀스 조회
         List<CarVO> carList = new ArrayList<>();    // 업체시퀀스로 카 리스트 조회
         String[] companyItems = {};                 // 업체시퀀스로 업체아이템 조회
 
         model.addAttribute("companySeq", companySeq);
         model.addAttribute("carList", carList);
         model.addAttribute("companyItems", companyItems);
-        model.addAttribute("estimate", estimateHistoryVO);
+        model.addAttribute("estimate", estimate);
         return "premiumRent/estimateSubmitInsert";
     }
